Use data with hex string
