+++ src/xcode/ENA/ENA/Source/Scenes/Events/Store/Model/__test__/TraceLocationTests.swift
@@ -207,23 +207,6 @@ class TraceLocationTests: CWATestCase {
 	}
 }
 
-private extension String {
-	func dataWithHexString() -> Data {
-		var hex = self
-		var data = Data()
-		while !hex.isEmpty {
-			let subIndex = hex.index(hex.startIndex, offsetBy: 2)
-			let c = String(hex[..<subIndex])
-			hex = String(hex[subIndex...])
-			var ch: UInt32 = 0
-			Scanner(string: c).scanHexInt32(&ch)
-			var char = UInt8(ch)
-			data.append(&char, count: 1)
-		}
-		return data
-	}
-}
-
 private extension Data {
 	func hexEncodedString() -> String {
 		let format = "%02hhx"
